+++ RangeHighlight/Integrations.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright 2020 Jamie Taylor
 using System;
+using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using StardewModdingAPI;
 
@@ -10,6 +11,7 @@ namespace RangeHighlight {
             this.theMod = theMod;
             IntegratePrismaticTools();
             IntegrateBetterJunimos();
+            IntegrateBetterSprinklers();
         }
 
         private void IntegratePrismaticTools() {
@@ -27,6 +29,35 @@ namespace RangeHighlight {
                 theMod.Monitor.Log($"ignoring nonsense value {r} from Better Junimos for Junimo Hut radius", LogLevel.Info);
             }
         }
+        private void IntegrateBetterSprinklers() {
+            IBetterSprinklersApi api = theMod.helper.ModRegistry.GetApi<IBetterSprinklersApi>("Speeder.BetterSprinklers");
+            if (api == null) return;
+            theMod.api.RemoveItemRangeHighlighter("jltaylor-us.RangeHighlight/sprinkler");
+            theMod.api.AddItemRangeHighlighter("jltaylor-us.RangeHighlight/better-sprinkler",
+                theMod.config.ShowSprinklerRangeKey,
+                theMod.config.ShowOtherSprinklersWhenHoldingSprinkler,
+                (item, itemID, itemName) => {
+                    Vector2[] tiles;
+                    if (api.GetSprinklerCoverage().TryGetValue(itemID, out tiles)) {
+                        return new Tuple<Color, bool[,]>(theMod.config.SprinklerRangeTint, PointsToMask(tiles));
+                    } else {
+                        return null;
+                    }
+                });
+        }
+        private bool[,] PointsToMask(Vector2[] points) {
+            int maxX = 0;
+            int maxY = 0;
+            foreach (var point in points) {
+                maxX = Math.Max(maxX, Math.Abs((int)point.X));
+                maxY = Math.Max(maxY, Math.Abs((int)point.Y));
+            }
+            bool[,] result = new bool[maxX * 2 + 1, maxY * 2 + 1];
+            foreach (var point in points) {
+                result[(int)point.X + maxX, (int)point.Y + maxY] = true;
+            }
+            return result;
+        }
     }
 
     public interface IPrismaticToolsAPI {
@@ -37,4 +68,10 @@ namespace RangeHighlight {
     public interface IBetterJunimosAPI {
         int GetJunimoHutMaxRadius();
     }
+
+    public interface IBetterSprinklersApi {
+        int GetMaxGridSize();
+        IDictionary<int, Vector2[]> GetSprinklerCoverage();
+    }
+
 }